# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Basic User Management API
  version: 1.0.0
servers:
  - url: http://localhost:3000
paths:
  /api/auth/register:
    post:
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../shared/schema/user.schema.json#/$defs/UserInsert
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: ../shared/schema/user.schema.json#/$defs/User
                required: [token, user]
        "400":
          description: Bad Request
  /api/auth/login:
    post:
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
              required: [email, password]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: ../shared/schema/user.schema.json#/$defs/User
                required: [token, user]
        "401":
          description: Unauthorized
  /api/me:
    get:
      operationId: me
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: ../shared/schema/user.schema.json#/$defs/User
                required: [user]
        "401":
          description: Unauthorized
  /api/users:
    get:
      operationId: listUsers
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: ../shared/schema/user.schema.json#/$defs/User
                required: [users]
    post:
      operationId: addUser
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../shared/schema/user.schema.json#/$defs/UserInsert
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: ../shared/schema/user.schema.json#/$defs/User
                required: [user]
        "400":
          description: Bad Request
  /api/users/{id}:
    patch:
      operationId: updateUser
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../shared/schema/user.schema.json#/$defs/UserUpdate
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: ../shared/schema/user.schema.json#/$defs/User
                required: [user]
  /api/auth/password/reset-request:
    post:
      operationId: passwordResetRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../shared/schema/user.schema.json#/$defs/PasswordResetRequest
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  token:
                    type: string
                required: [ok]
  /api/auth/password/reset-confirm:
    post:
      operationId: passwordResetConfirm
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../shared/schema/user.schema.json#/$defs/PasswordResetConfirm
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required: [ok]
    delete:
      operationId: deleteUser
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: ../shared/schema/user.schema.json#/$defs/User
                required: [user]
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT